Fix all clippy warnings and add user memory validation
- Resolved 589 clippy warnings across codebase
- Added kernel-level lint configurations for kernel-specific patterns
- Implemented user memory validation functions (probe_for_read/write)
- Improved NtReadVirtualMemory and NtWriteVirtualMemory with validation
- Replaced manual ASCII case comparisons with std methods
- Renamed from_str methods to new_from to avoid trait confusion
- Fixed various code quality issues (clamp, transmute annotations, etc.)

+++ kernel/src/arch/x86_64/context.rs
@@ -119,7 +119,7 @@ pub unsafe fn setup_initial_context(
 
     // Push return address (entry point)
     sp -= 8;
-    *(sp as *mut u64) = entry_point as u64;
+    *(sp as *mut u64) = entry_point as usize as u64;
 
     // Push callee-saved registers (all zero for new thread)
     sp -= 8;
